+++ evenement-edit.php
@@ -471,7 +471,7 @@ if (isset($_POST['formulaire']) && $_POST['formulaire'] === 'ok' )
 			{
 				$req_id = $connector->getInsertId();
 
-				$_SESSION['evenement-edit_flash_msg'] = "L'événement a été créé. <a href='/agenda.php?courant=".$champs['dateEvenement']."#event-".$req_id."'>Voir dans l'agenda</a>";
+				$_SESSION['evenement-edit_flash_msg'] = "L'événement a été créé. <a href='/evenement-agenda.php?courant=".$champs['dateEvenement']."#event-".$req_id."'>Voir dans l'agenda</a>";
 
                 if (!isset($_SESSION['Sgroupe']))
                 {
@@ -653,7 +653,7 @@ if (isset($_POST['formulaire']) && $_POST['formulaire'] === 'ok' )
                     $mailer->toUser($champs['user_email'], $subject, $contenu_message);
                 }
 
-                $_SESSION['evenement-edit_flash_msg'] = "L'événement a été modifié.$confirmation_flash_msg<br><a href='/agenda.php?courant=".$champs['dateEvenement']."#event-".$req_id."'>Voir dans l'agenda</a>";
+                $_SESSION['evenement-edit_flash_msg'] = "L'événement a été modifié.$confirmation_flash_msg<br><a href='/evenement-agenda.php?courant=".$champs['dateEvenement']."#event-".$req_id."'>Voir dans l'agenda</a>";
 
 				$get['action'] = 'editer';
 
@@ -894,7 +894,7 @@ if ($verif->nbErreurs() > 0)
         <?php } ?>
         <p>Veillez svp à ce que votre événement</p>
         <ul style="line-height:1.2em">
-            <li style="margin:6px 2px;">n’est pas déjà présent dans l’<a href="/agenda.php" target="_blank">agenda</a></li>
+            <li style="margin:6px 2px;">n’est pas déjà présent dans l’<a href="/evenement-agenda.php" target="_blank">agenda</a></li>
             <li style="margin:6px 2px;">respecte notre <a href="/articles/charte-editoriale.php" target="_blank">charte&nbsp;éditoriale</a></li>
         </ul>
     </div>